change init_flock and add_boid functions
both functions have been changed to keep flexible, as previously boid values were hardcoded in.
struct flock now stores the Geometry that each boid will use to render itself.
add_boid function now takes x and y values to initialize it's position to, and no longer takes a pointer to Geometry, as this info is now it it's flock.